Move JsonDocument asserter to better namespace
