+++ tests/ElasticSearch/Organizer/OrganizerJsonDocumentTransformerTest.php
@@ -5,7 +5,7 @@ namespace CultuurNet\UDB3\Search\ElasticSearch\Organizer;
 use CultuurNet\UDB3\ReadModel\JsonDocument;
 use CultuurNet\UDB3\Search\ElasticSearch\PathEndIdUrlParser;
 use CultuurNet\UDB3\Search\ElasticSearch\SimpleArrayLogger;
-use CultuurNet\UDB3\Search\Json\AssertJsonDocument;
+use CultuurNet\UDB3\Search\JsonDocument\AssertJsonDocument;
 use PHPUnit\Framework\TestCase;
 
 class OrganizerJsonDocumentTransformerTest extends TestCase
