+++ tests/JsonDocument/PassThroughJsonDocumentTransformerTest.php
@@ -3,7 +3,6 @@
 namespace CultuurNet\UDB3\Search\JsonDocument;
 
 use CultuurNet\UDB3\ReadModel\JsonDocument;
-use CultuurNet\UDB3\Search\Json\AssertJsonDocument;
 use PHPUnit\Framework\TestCase;
 
 class PassThroughJsonDocumentTransformerTest extends TestCase
